Create tables, schemas, etc. Favour more data less logic
- Add cluster configuration parameters, to control
mnesia database nodes

- Simplify data carried around by schedule when
running. Most of the data will now be read
from db when needed. All we need when alive
and running is the Timer ref and name

- add sane defaults for #schedule{} record when
creating, so we do not have to specify everything

- fleshout schedule_store, to support proper
db system initialization and stopping or destruction

- implement schedule add to db. Not yet properly tested

- really simplify the schedule fsm. Actually, what I've
decided is to only start the fsm for enabled processes.

When in any other state, the fsm will not be running.

- Comment out lots of code for now, to get things properly
hooked up. We'll flesh them out again later.

- Also, simplify the scheduler and couple it properly with the
new schedules. scheduler no more needs to keep schedule map
in its state data

- couple scheduler with schedule_store, so they work well
in tandem

- Add node name to the unittests, so db tests work well

- update all unittests to pass